work on mutation tests, bugfix
